Fastapi/requirements.txt
fastapi
uvicorn
numpy
tensorflow
pillow
python-multipart
# scikit-learn is often needed for tf legacy/other reasons, but strict reqs are above
